Cmake lear Zhoholiev Pavlo 26.12.24 -make static, interface, public
