+++ home_work_custom_shared_ptr/src/custom_shared_ptr.cpp
@@ -1,0 +1,10 @@
+//
+// Created by flyon21 on 26.12.24.
+//
+#include "../include/custom_shared_ptr.hpp"
+
+namespace custom_shared_ptr {
+    const char* getVersion() {
+        return "0.1.0";
+    }
+}
